+++ Paramount.Betterclassifieds.Business/Events/EventBookingContext.cs
@@ -1,10 +1,25 @@
 ﻿namespace Paramount.Betterclassifieds.Business.Events
 {
+    public interface IEventBookingContext
+    {
+        int? EventId { get; set; }
+        int? EventBookingId { get; set; }
+        string EventBookingPaymentReference { get; set; }
+        string[] EmailGuestList { get; set; }
+        string Purchaser { get; set; }
+        void Clear();
+    }
+
     /// <summary>
     /// Object used for session storage when booking tickets to an event
     /// </summary>
-    public class EventBookingContext
+    public class EventBookingContext : IEventBookingContext
     {
+        public EventBookingContext()
+        {
+            EmailGuestList = new string[0];
+        }
+
         public int? EventId { get; set; }
         public int? EventBookingId { get; set; }
         public string EventBookingPaymentReference { get; set; }
@@ -16,7 +31,7 @@
             EventId = null;
             EventBookingId = null;
             EventBookingPaymentReference = null;
-            EmailGuestList = null;
+            EmailGuestList = new string[0];
             Purchaser = null;
         }
     }